feat: move all configuration classes in new files
