+++ src/fault_injector/fault_configurations/fault_configuration.py
@@ -1,0 +1,24 @@
+from src.component import Component
+from src.base_model import BaseModel
+import marshmallow as marsh
+from peewee import IntegerField, TextField
+
+class FaultConfiguration(BaseModel):
+    """Class that contains the attributes of the Fault class"""
+
+    class Schema(BaseModel.Schema):
+        """Schema for the FaultConfiguration"""
+
+        start_tick = marsh.fields.Integer(required=True)
+        end_tick = marsh.fields.Integer(required=True)
+        description = marsh.fields.String()
+
+        def _make(self, data: dict) -> "FaultConfiguration":
+            return FaultConfiguration(**data)
+
+    start_tick = IntegerField(null=False)
+    end_tick = IntegerField(null=False)
+
+    component: Component = None
+    # - affected_element_ID: int = None // has to be implemented in subclasses
+    description = TextField(default="injected Fault")
