+++ src/fault_injector/fault_types/platform_blocked_fault.py
@@ -1,8 +1,5 @@
-import marshmallow as marsh
-from peewee import TextField
-
 from src.component import Component
-from src.fault_injector.fault_types.fault import Fault, FaultConfiguration
+from src.fault_injector.fault_types.fault import Fault
 
 
 class PlatformBlockedFault(Fault):
@@ -29,15 +26,4 @@ class PlatformBlockedFault(Fault):
         raise NotImplementedError()
 
 
-class PlatformBlockedFaultConfiguration(FaultConfiguration):
-    """Class that contains the attributes of the PlatformBlockedFault class"""
 
-    class Schema(FaultConfiguration.Schema):
-        """Schema for PlatformBlockedFaultConfiguration"""
-
-        affected_element_id = marsh.fields.String(required=True)
-
-        def _make(self, data: dict) -> "PlatformBlockedFaultConfiguration":
-            return PlatformBlockedFaultConfiguration(**data)
-
-    affected_element_id = TextField(null=False)
