+++ src/fault_injector/fault_types/track_blocked_fault.py
@@ -1,8 +1,5 @@
-import marshmallow as marsh
-from peewee import TextField
-
 from src.component import Component
-from src.fault_injector.fault_types.fault import Fault, FaultConfiguration
+from src.fault_injector.fault_types.fault import Fault
 
 
 class TrackBlockedFault(Fault):
@@ -29,15 +26,4 @@ class TrackBlockedFault(Fault):
         raise NotImplementedError()
 
 
-class TrackBlockedFaultConfiguration(FaultConfiguration):
-    """Class that contains the attributes of the TrackBlockedFault class"""
 
-    class Schema(FaultConfiguration.Schema):
-        """Schema for TrackBlockedFaultConfiguration"""
-
-        affected_element_id = marsh.fields.String()
-
-        def _make(self, data: dict) -> "TrackBlockedFaultConfiguration":
-            return TrackBlockedFaultConfiguration(**data)
-
-    affected_element_id = TextField()
