+++ src/fault_injector/fault_types/track_speed_limit_fault.py
@@ -1,8 +1,5 @@
-import marshmallow as marsh
-from peewee import IntegerField, TextField
-
 from src.component import Component
-from src.fault_injector.fault_types.fault import Fault, FaultConfiguration
+from src.fault_injector.fault_types.fault import Fault
 
 
 class TrackSpeedLimitFault(Fault):
@@ -31,17 +28,4 @@ class TrackSpeedLimitFault(Fault):
         raise NotImplementedError()
 
 
-class TrackSpeedLimitFaultConfiguration(FaultConfiguration):
-    """Class that contains the attributes of the TrackSpeedLimitFault class"""
 
-    class Schema(FaultConfiguration.Schema):
-        """Schema for TrackSpeedLimitFaultConfiguration"""
-
-        affected_element_id = marsh.fields.String()
-        new_speed_limit = marsh.fields.Integer(required=True)
-
-        def _make(self, data: dict) -> "TrackSpeedLimitFaultConfiguration":
-            return TrackSpeedLimitFaultConfiguration(**data)
-
-    affected_element_id = TextField()
-    new_speed_limit = IntegerField(null=False)
