+++ src/fault_injector/fault_types/train_cancelled_fault.py
@@ -1,8 +1,5 @@
-import marshmallow as marsh
-from peewee import TextField
-
 from src.component import Component
-from src.fault_injector.fault_types.fault import Fault, FaultConfiguration
+from src.fault_injector.fault_types.fault import Fault
 
 
 class TrainCancelledFault(Fault):
@@ -27,15 +24,4 @@ class TrainCancelledFault(Fault):
         raise NotImplementedError()
 
 
-class TrainCancelledFaultConfiguration(FaultConfiguration):
-    """Class that contains the attributes of the TrainCancelledFault class"""
 
-    class Schema(FaultConfiguration.Schema):
-        """Schema for TrainCancelledFaultConfiguration"""
-
-        affected_element_id = marsh.fields.String()
-
-        def _make(self, data: dict) -> "TrainCancelledFaultConfiguration":
-            return TrainCancelledFaultConfiguration(**data)
-
-    affected_element_id = TextField()
