+++ src/fault_injector/fault_types/train_prio_fault.py
@@ -1,8 +1,5 @@
-import marshmallow as marsh
-from peewee import IntegerField, TextField
-
 from src.component import Component
-from src.fault_injector.fault_types.fault import Fault, FaultConfiguration
+from src.fault_injector.fault_types.fault import Fault
 
 
 class TrainPrioFault(Fault):
@@ -31,17 +28,4 @@ class TrainPrioFault(Fault):
         raise NotImplementedError()
 
 
-class TrainPrioFaultConfiguration(FaultConfiguration):
-    """Class that contains the attributes of the TrainPrioFault class"""
 
-    class Schema(FaultConfiguration.Schema):
-        """Schema for TrainPrioFaultConfiguration"""
-
-        affected_element_id = marsh.fields.String()
-        new_prio = marsh.fields.Integer(required=True)
-
-        def _make(self, data: dict) -> "TrainPrioFaultConfiguration":
-            return TrainPrioFaultConfiguration(**data)
-
-    affected_element_id = TextField()
-    new_prio = IntegerField(null=False)
