+++ src/fault_injector/fault_types/train_speed_fault.py
@@ -1,8 +1,5 @@
-import marshmallow as marsh
-from peewee import TextField
-
 from src.component import Component
-from src.fault_injector.fault_types.fault import Fault, FaultConfiguration
+from src.fault_injector.fault_types.fault import Fault
 
 
 class TrainSpeedFault(Fault):
@@ -31,15 +28,4 @@ class TrainSpeedFault(Fault):
         raise NotImplementedError()
 
 
-class TrainSpeedFaultConfiguration(FaultConfiguration):
-    """Class that contains the configuration attributes of the TrainSpeedFault class"""
 
-    class Schema(FaultConfiguration.Schema):
-        """Schema for TrainSpeedFaultConfiguration"""
-
-        affected_element_id = marsh.fields.String(required=True)
-
-        def _make(self, data: dict) -> "TrainSpeedFaultConfiguration":
-            return TrainSpeedFaultConfiguration(**data)
-
-    affected_element_id = TextField(null=False)
